src/app/ops-assistant/requirements.txt
# LangChain 核心依赖 - 最新版本 (2026-02)
langchain>=1.2.10
langchain-core>=1.2.14
langchain-openai>=1.1.10
langchain-community>=0.4.1
langgraph>=1.0.9

# 环境变量管理
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0

# HTTP 客户端
httpx>=0.27.0
